+++ src/main/java/org/geojson/Feature.java
@@ -1,11 +1,7 @@
 package org.geojson;
 
-import java.util.HashMap;
-import java.util.Map;
-
 public class Feature extends GeoJsonObject {
 
-	private Map<String, Object> properties = new HashMap<>();
 	private GeoJsonObject geometry;
 	private String id;
 
@@ -15,23 +11,6 @@ public class Feature extends GeoJsonObject {
 
 	public void setGeometry(GeoJsonObject geometry) {
 		this.geometry = geometry;
-	}
-
-	public void setProperty(String key, Object value) {
-		properties.put(key, value);
-	}
-
-	@SuppressWarnings("unchecked")
-	public <T> T getProperty(String key) {
-		return (T)properties.get(key);
-	}
-
-	public Map<String, Object> getProperties() {
-		return properties;
-	}
-
-	public void setProperties(Map<String, Object> properties) {
-		this.properties = properties;
 	}
 
 	public String getId() {
