+++ src/main/java/org/geojson/GeoJsonObject.java
@@ -7,6 +7,9 @@ import com.fasterxml.jackson.annotation.JsonSubTypes.Type;
 import com.fasterxml.jackson.annotation.JsonTypeInfo;
 import com.fasterxml.jackson.annotation.JsonTypeInfo.Id;
 
+import java.util.HashMap;
+import java.util.Map;
+
 @JsonTypeInfo(property = "type", use = Id.NAME)
 @JsonSubTypes({ @Type(Feature.class), @Type(Polygon.class), @Type(MultiPolygon.class), @Type(FeatureCollection.class),
 		@Type(Point.class), @Type(MultiPoint.class), @Type(MultiLineString.class), @Type(LineString.class) })
@@ -15,6 +18,9 @@ public abstract class GeoJsonObject {
 
 	private Crs crs;
 	private double[] bbox;
+
+  @JsonInclude(Include.NON_EMPTY)
+	private Map<String, Object> properties = new HashMap<>();
 
 	public Crs getCrs() {
 		return crs;
@@ -30,5 +36,22 @@ public abstract class GeoJsonObject {
 
 	public void setBbox(double[] bbox) {
 		this.bbox = bbox;
+	}
+	
+  public void setProperty(String key, Object value) {
+		properties.put(key, value);
+	}
+
+	@SuppressWarnings("unchecked")
+	public <T> T getProperty(String key) {
+		return (T)properties.get(key);
+	}
+
+	public Map<String, Object> getProperties() {
+		return properties;
+	}
+
+	public void setProperties(Map<String, Object> properties) {
+		this.properties = properties;
 	}
 }
